+++ src/lib.rs
@@ -54,6 +54,8 @@ pub use self::event::Key;
 pub use self::event::MetaEvent;
 pub use self::event::OptionChain;
 pub use self::event::UiEvent;
+pub use self::event::UnhandledEvent;
+pub use self::event::UnhandledEvents;
 pub use self::handleable::Handleable;
 pub use self::object::Object;
 pub use self::renderable::Renderable;